+++ packages/startupjs/bundler/react-native.config.js
@@ -1,1 +1,0 @@
-export { default } from '@startupjs/bundler/react-native.config.js'