+++ packages/startupjs/bundler/webpack.web.config.js
@@ -1,1 +1,0 @@
-export { default } from '@startupjs/bundler/webpack.web.config.js'