+++ packages/startupjs/templates/simple/metro.config.cjs
@@ -1,0 +1,5 @@
+const config = require('startupjs/bundler/metro.config.cjs')
+
+// Override default metro config here.
+
+module.exports = config