+++ packages/startupjs/templates/simple/webpack.web.config.js
@@ -1,6 +1,0 @@
-import getConfig from 'startupjs/bundler/webpack.web.config.cjs'
-
-export default getConfig(undefined, {
-  forceCompileModules: [],
-  mode: 'react-native'
-})