+++ styleguide/metro.config.js
@@ -1,9 +1,3 @@
-import config from 'startupjs/bundler/metro.config.js'
-import path from 'path'
-
-config.watchFolders = [
-  path.resolve(__dirname, '../'),
-  path.resolve(__dirname, '../node_modules')
-]
-
-export default config
+// FIXME: Metro does not provide way to use cjs config.
+// https://github.com/facebook/metro/blob/master/packages/metro-config/src/loadConfig.js#L49
+module.exports = require('./metro.config.cjs')