+++ styleguide/react-native.config.js
@@ -1,6 +1,0 @@
-import getConfig from 'startupjs/bundler/react-native.config.js'
-
-export default getConfig({
-  project: {},
-  assets: []
-})