+++ styleguide/webpack.web.config.cjs
@@ -1,0 +1,7 @@
+const getConfig = require('startupjs/bundler/webpack.web.config.cjs')
+
+module.exports = getConfig(undefined, {
+  forceCompileModules: [],
+  alias: {},
+  mode: 'react-native'
+})